Remove DateOfBirth field from Edit form
The code block handling the "DateOfBirth" form field has been removed from the `Edit.cshtml` file. This includes the label, input, and validation span for the "DateOfBirth" field. The rest of the form remains unchanged. This may have been added by accident when creating this view. Removed the "DateOfBirth" field from the edit form as it's not relevant to appointment editing.